refactor test folder
